Use the highlight.js for client side code highlighting.
Code markup must be highlighted everytime JIRA updates content on an
issue page. So we bind the highlight action to
JIRA.Events.NEW_CONTENT_ADDED which triggers when inline editing of a
comment occurs (for example).

We use the 'showWikiInput' JQuery event (triggered by JIRA when markup
is rendered for a preview) to apply highlighting to previews.

The former event is part of JIRA's JavaScript API [1] the latter is not, so may
change between JIRA versions.

[1] https://developer.atlassian.com/jiradev/latest-updates/developer-changes-for-older-jira-versions/preparing-for-jira-5-0/custom-fields-that-use-css-or-javascript-web-resources-in-jira-5-0
https://jira.int.corefiling.com/browse/REVIKI-587